requirements.txt
pytest
pylint
ipdb
jupyter
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
nltk
google-cloud-storage
pyarrow
blob
os
colorama

# API
fastapi
pytz
uvicorn

# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest
